+++ lib/bookingsync_application.rb
@@ -1,8 +1,8 @@
-require "bookingsync"
-require "synced"
-require "dotenv-rails"
-require "jsonapi-resources"
-require "bookingsync_application/engine"
+require 'bookingsync'
+require 'synced'
+require 'dotenv-rails'
+require 'jsonapi-resources'
+require 'bookingsync_application/engine'
 require 'bookingsync_application/common_base_controller'
 
 module BookingsyncApplication