+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # master
 
+* bumped jsonapi resources to 0.0.16
+* added rubocop and initial cleanup
+
 # 0.1.3 (2015-02-20)
 
 * require dotenv-rails and not dotenv